Give better semantic name to Builder and Eventifier
They are targeting a full XML document, which is now made clear by using
proper naming.
