+++ documentation/docs/xml/index.md
@@ -68,30 +68,30 @@ normalized.compile.toList.unsafeRunSync()
 
 To handle XML DOM, you can use the types and pipes available in the `fs2.data.xml.dom` package.
 
-XML DOM can be built if you provide an implicit [`Builder[Node]`][builder-api] to the `documents` pipe. The `Builder[Node]` typeclass describes how XML DOM of type `Node` are built from an XML event stream.
+XML DOM can be built if you provide an implicit [`DocumentBuilder[Doc]`][builder-api] to the `documents` pipe. The `DocumentBuilder[Doc]` typeclass describes how XML DOM of type `Doc` are built from an XML event stream.
 
 ```scala mdoc:compile-only
 import dom._
 
-trait SomeNodeType
+trait SomeDocType
 
-implicit val builder: Builder[SomeNodeType] = ???
-stream.through(documents[IO, SomeNodeType])
+implicit val builder: DocumentBuilder[SomeDocType] = ???
+stream.through(documents[IO, SomeDocType])
 ```
 
-Conversely, the pipe transforming a stream of `Node`s into a stream of XML events is called `eventify` and requires an implicit [`Eventifier[Node]`][eventifier-api] in scope.
+Conversely, the pipe transforming a stream of `Doc`s into a stream of XML events is called `eventify` and requires an implicit [`DocumentEventifier[Doc]`][eventifier-api] in scope.
 
 ```scala mdoc:compile-only
 import dom._
 
-trait SomeNodeType
+trait SomeDocType
 
-implicit val builder: Builder[SomeNodeType] = ???
-implicit val eventifier: Eventifier[SomeNodeType] = ???
+implicit val builder: DocumentBuilder[SomeDocType] = ???
+implicit val eventifier: DocumentEventifier[SomeDocType] = ???
 
-stream.through(documents[IO, SomeNodeType])
-      .through(eventify[IO, SomeNodeType])
+stream.through(documents[IO, SomeDocType])
+      .through(eventify[IO, SomeDocType])
 ```
 
-[builder-api]: /api/fs2/data/xml/dom/Builder.html
-[eventifier-api]: /api/fs2/data/xml/dome/Eventifier.html
+[builder-api]: /api/fs2/data/xml/dom/DocumentBuilder.html
+[eventifier-api]: /api/fs2/data/xml/dom/DocumentEventifier.html
