+++ xml/src/main/scala/fs2/data/xml/dom/DocumentBuilder.scala
@@ -18,8 +18,25 @@ package data
 package xml
 package dom
 
-trait Eventifier[Node] {
+trait DocumentBuilder[Document] {
 
-  def eventify(node: Node): Stream[Pure, XmlEvent]
+  type Content
+  type Misc <: Content
+  type Elem <: Content
+
+  def makeDocument(version: Option[String],
+                   encoding: Option[String],
+                   standalone: Option[Boolean],
+                   doctype: Option[XmlEvent.XmlDoctype],
+                   prolog: List[Misc],
+                   root: Elem): Document
+
+  def makeComment(content: String): Option[Misc]
+
+  def makeText(texty: XmlEvent.XmlTexty): Content
+
+  def makeElement(name: QName, attributes: List[Attr], isEmpty: Boolean, children: List[Content]): Elem
+
+  def makePI(target: String, content: String): Misc
 
 }
